src/test/java/exercises/introduction/StdinAndStdoutIITest.java
package exercises.introduction;

import org.junit.After;
import org.junit.Before;
import org.junit.Test;

import java.io.ByteArrayInputStream;
import java.io.ByteArrayOutputStream;
import java.io.PrintStream;

import static java.lang.System.in;
import static java.lang.System.out;
import static java.lang.System.setIn;
import static java.lang.System.setOut;
import static org.assertj.core.api.Assertions.assertThat;

public class StdinAndStdoutIITest {
    private static final String INPUT = "42\n3.1415\nWelcome to HackerRank's Java tutorials!!\n";
    private static final String OUTPUT = "String: Welcome to HackerRank's Java tutorials!!\nDouble: 3.1415\nInt: 42\n";

    private final ByteArrayOutputStream outContent = new ByteArrayOutputStream();
    private final ByteArrayInputStream inContent = new ByteArrayInputStream(INPUT.getBytes());

    @Before
    public void setUp() {
        setOut(new PrintStream(outContent));
        setIn(inContent);
    }

    @Test
    public void shouldReadAndPrintStringDoubleAndInteger()  {
        StdinAndStdoutII exercise = new StdinAndStdoutII();

        exercise.solve();

        assertThat(outContent.toString()).isEqualTo(OUTPUT);
    }

    @After
    public void tearDown() {
        setOut(out);
        setIn(in);
    }
}